requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.6
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
limits==3.2.0
MarkupSafe==2.1.2
msgpack==1.0.4
openapi-codec==1.3.2
packaging==23.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing-extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
zope.interface==5.5.2
